Remove blanket impl of TrivialDrop for Copy types, constrain generic parameters
Summary:
The blanket impl is convenient, but it's not right. If we have a tuple `(T1, T2)`, where T1 and T2 are both TrivialDrop, but one or both are not Copy, then the tuple is not TrivialDrop. I don't think we can write `impl<T1: TrivialDrop, T2: TrivialDrop> TrivialDrop for (T1, T2) {}` as long as we have the blanket impl, so we must remove it if we would like tuples like this to implement the marker trait.

Additionally, there's an oversight in the implementations of TrivialDrop for generic oxidized_by_ref types like those defined in aast.rs. Each should only implement TrivialDrop if all generic parameters implement it. This diff adds those constraints.

Differential Revision: D22364272

fbshipit-source-id: 90a9b0d56c1e13d660ac43e6ffd3b727455241d4