Write to files as inline commands in tests
Summary:
The current model is that files are contained in the repo in the form `new_file.php`, `new_file.php.new1`, etc. This makes it hard to figure out what the test is doing in the aggregate.

This commit changes it so that the file writes are embedded into the test case itself.

Reviewed By: ljw1004

Differential Revision: D21969497

fbshipit-source-id: 042c64d4d63f8ebf2b5a96c3fd0b02522530745e