Modify ExternCompiler hook interface
Summary:
Modify the ExternCompiler hook interface to return UnitEmitters
directly instead of HHAS. This gives the implementation of the hook
more freedom (it can generate UnitEmitters directly instead of forcing
HHAS generation). Instead of one compile function being provided to
the hook, two are. One generates HHAS (as before), and the second will
convert HHAS to an UnitEmitter. This gives the hook maximum
flexibility and does not require duplication of the assembler
logic. Simplify the interface a bit and remove some un-used logging.

Reviewed By: subsistence, paulbiss

Differential Revision: D22428830

fbshipit-source-id: 2dbcf2621c7f94de19221eaa172982a5edce8597