track server times for request-handling
Summary:
It used to be that we tracked only a single time for how a server handled a request - namely, the time that the server entered `ServerMain.handle_connection_`.

This diff changes it to gather a load more timing information about all stages of how the server handled the request.

It gathers this data inside Connection_tracker.t, because I want a single strongly-typed place that can gather+correlate telemetry about the entire end-to-end response. Only after the response gets sent back to the client will our telemetry be complete.

Reviewed By: CatherineGasnier

Differential Revision: D22417859

fbshipit-source-id: 221d2208b9f02915d42711693db5530b357ebb62