Connection_tracker.t handed back to client
Summary:
My goal is to have a single consolidated way to track how we handled a request.

We used to have just a rudimentary form -- when client made a request, then the server sent back a single float "start_server_handle_time" for when it entered `ServerMain.handle_connection_`. The client took this as the time that the server was unblocked enough to handle its request.

With this diff, the server sends back not a float but rather the Connection_tracker.t object that it received from the monitor. It also stores that "start_server_handle_time" inside the Connection_tracker.t. This is the start of a platform by which the client will be able to get much richer end-to-end telemetry about how the request was handled in all phases of its work.

Reviewed By: 2BitSalute

Differential Revision: D22417771

fbshipit-source-id: 2951e81ee69edf01c8be7cc7fc0cdabc9d0fac97