trivial connection logging
Summary:
I'm trying to track down connection issues between client/server/monitor.

I recently introduced a "monitor connection id" on the client side so that I can correlate each interaction with from the client to the monitor. This diff adds logging of that id to a few places that were missing it.

Reviewed By: arxanas

Differential Revision: D22384367

fbshipit-source-id: 79c4ad07f353bcb68fc2179cd56f6c5eb5e840ec