Connection_tracker.t in client
Summary:
My goal is and end-to-end tracking of connection requests, in a Connection_tracker.t record.

This diff switches the client to use Connection_tracker.t. It only uses it to replace the "ID" tracking that client was already doing. It doesn't yet do anything further with it -- doesn't pass it on to monitor/server, nor record timings. That will happen later in the stack.

Reviewed By: 2BitSalute

Differential Revision: D22417478

fbshipit-source-id: 769d89962396006fe578610173d9419566f003c9