Connection_tracker.t handed off to monitor
Summary:
My goal is to have a single place to gather telemetry across the whole client-monitor-server connection, Connection_tracker.t

This diff changes the client to hand over its Connection_tracker.t object to the monitor within the existing (badly named) HANDOFF_TO_SERVER message. It also changes the monitor to include the request's unique tracker-ID in its logs.

We don't yet send the tracker on to the server, nor record timings in it. That will come in a later diff.

Reviewed By: arxanas

Differential Revision: D22384444

fbshipit-source-id: d94531e94dfe0dfcfbeeceb926ba7563037eb16a