track client times for request-handling
Summary:
My goal is to have detailed end-to-end telemetry on how we handled a request.

This diff adds in telemetry from the client side.

Note: the model I've used for telemetry is to gather it into a `Connection_tracker.t` record, which is created at client, then handed off to monitor, then handed off to server, then handed back to client.

Where to store timings for things that happen inside the client after the tracker has been handed off to the monitor? -- I've stored these timings inside the `conn` record, so that when the tracker is eventually handed back from the server, then I can update it.

Reviewed By: CatherineGasnier

Differential Revision: D22417932

fbshipit-source-id: aaa3f6caa1fd290a6ec76c1a4d36d156d65b52b4