Don't call `Path.make` on changed files
Summary: We're currently calling `Path.make` on all changed-files that we've been notified about by the LSP file-watching subscription. This can cause unintentional Eden fetches. Switch to passing around the `string` path and not calling `Path.make` on it. (Eventually, we convert it into a `Relative_path.t` — but only once we've confirmed that it's a file we care about, i.e. a Hack file rooted in the correct repo.)

Reviewed By: ljw1004

Differential Revision: D22466465

fbshipit-source-id: 5fa3400b59f79a8ce29d7704e5383dd8dedc971d