Logging for HAM-safe intish casts
Summary:
Let's face it: we're not going to get rid of intish-cast. It's happening all over our large Hack codebase. However, we can make intish-cast HAM-safe, by expanding the builtins that do it for PHP arrays to also do it for Hack arrays.

For most builtins, expanding intish-cast to all array-likes is a no-op, because they only  intish-cast when setting to their result (which is statically known to be either a darray or a plain array). However, there are a few builtins that intish-cast with an input array - most notably, array_column - and for these builtins, expanding intish cast would change behavior.

To fix the issue, we'll log when we "miss" an intish-cast because the input is a Hack array. See new unit test for an example of how this logging works.

Reviewed By: ricklavoie

Differential Revision: D22472912

fbshipit-source-id: 6c5afee76ca73c06317bee4d96c0e659f4698210