Add another granularity level to ForbidDynamic* options
Summary: With `Eval.ForbidDynamicCallsWithAttr=true`, it's now possible to throw on a dynamic call to a function or method that hasn't been opted in, but still log a notice at the callsite for not using a pointer obtained from `HH\dynamic_fun()`, etc., by setting `Eval.ForbidDynamicCallsToFunc`, etc. to the value 2. To also throw when the callsite isn't using a pointer, set it to the value 3.

Reviewed By: ricklavoie

Differential Revision: D22427185

fbshipit-source-id: 43532faa58a7d7f9b114c4f424dc75dce35bdeb7