Get tests ready for string conv ban
Summary:
I'll just keep adding tests here until the next diff passes all tests.

General principles: if the behavior being tested is silly, like an explicit string cast of a dict, just remove it. If the test failure is just due to the common usage of formatting arguments as in "Current value: $x", then compute $text = HH\is_any_array($x) ? 'Array' : $x; and use that as the display value. This expression is HAM-safe so we won't need to revisit it.

Reviewed By: rodmk

Differential Revision: D22450648

fbshipit-source-id: 0a8ee26e58f297f7391682aa19dbcc73312ed0a2