Stop intish-casting in strtr
Summary:
strtr takes an input string and an array-like mapping from pattern -> replacement. For dicts, strtr works as you might expect. For PHP arrays, strtr's behavior is equivalent to silently dropping any intish-string keys from the array! Read on for why it has this weird behavior. We drop it because it's a Bad Idea.

We do a strange intish-cast when strtr is called with a string input and a darray mapping pattern strings to replacement strings. To implement this method, we first compute various rolling hashes of the patterns so we can efficiently prune the pattern checks; when the hash of some substring of the input matches a pattern, we do a full substring check.

The strange intish-cast is when we do a lookup of the input substring in the pattern darray. It triggers if there's a hash collision between a intish substring of the input and a non-intish pattern. However, as long as the patterns darray doesn't contain any intish-string keys, this intish-cast never affects the result.

If the patterns darray actually does contain any intish-string patterns, we'll do a lookup for the int key instead, fail, and never do the replacement. Thus, the behavior for darrays is equivalent to just dropping all intish-string patterns.

Reviewed By: dneiter

Differential Revision: D22497336

fbshipit-source-id: 9da960b34f4b5c3afebc208bf27c099a78104df7