Fix plain-array-observing tests
Summary:
A bunch of tests created plain arrays and observed logging behavior (e.g. no Hack Arr Compat logging for is_dict, or no provenance) that will go away if we produce darrays instead. Other tests observe comparison behavior, and yet other tests observe typehints. Finally, there are some special serialization intrinsics (serialize_keep_dvarrays) not exposed to Hack code yet that distinguish plain arrays and darrays.

It's safe to change all of these behaviors for the reasons described in D22365647. We drop the tests, or the lines in the tests, that test them.

There is one behavior I've found that actually changes. SimpleXMLElement relational comparison passes through to array comparison on the underlying array. There's logging for this behavior, which I'm going to turn on; it shouldn't be hard to ban.

Reviewed By: ricklavoie

Differential Revision: D22498665

fbshipit-source-id: 7c9111fca93253c05e0f4fac4767f6bbc4f2ae0a