Delete unused comparison functions
Summary:
There are three separate implementations of array-like to array-like comparison: in tv-comparisons, in type-array, and in array-data.cpp. Having so many cases makes it hard to keep logging in sync between them all. That's not good, because I'm pretty sure our logging for the final HAM switch is incomplete and/or wrong.

Turns out we can kill ALL the comparison functions in type-array, and ArrayData::compare, and the weak case of ArrayData::equal. Yay!

Reviewed By: dneiter

Differential Revision: D22503312

fbshipit-source-id: 061e4d66414004fb29c496686a487f3da0622605