Add logging for comparison behaviors
Summary:
I've found a way that we can trigger a comparison between two plain arrays: SimpleXML's custom ObjectData comparator. Just to be safe, let's add notices for the observable behavior differences that D22365647 would cause. The typehints are fine, as they truly are only executed on Hack-visible values, but it may be possible for obscure C++ code paths to execute Eq/Same on a plain array + darray, or Compare on two plain arrays, both of which will change behavior.

Eq/Same on two plain arrays is fine; it'll return the same result before and after we replace these arrays with darrays. Likewise, Compare on a plain array + darray is fine; it already throws.

Reviewed By: ricklavoie

Differential Revision: D22498748

fbshipit-source-id: 178bfac5e04f53fb4a756eae30a135d2ca9817f0