Always use darray/varray in error messages
Summary:
We use darray/varray in some type errors (those connected directly to typehints), but not in others. Let's be consistent. This diff will change a bunch of builtin and member-operation error messages. I'll fix the tests as CI identifies them.

There's one place where we use getDataTypeString for reasons other than formatting an error message: in gettype. We continue to return "array" for dvarrays at this point.

Reviewed By: ricklavoie

Differential Revision: D22511813

fbshipit-source-id: 215f95ffb93777eeeba02fa348e078cecc1415ae