Adds code from context to error message, prefixed by line number and a column " |".
Summary:
This adds code from the context of an error message to the highlighted error formatter, prefixing with a line number and "|". The size of the column is currently only based on the length of the largest number referenced in the error code contexts.

One function, "format_context_highlighted", uses a couple helper functions initially written for just the context highlighter; in future diffs when I need added abilities, I will implement my own versions of those instead.

Reviewed By: Wilfred

Differential Revision: D22448264

fbshipit-source-id: c1bf4eafe37d0886b04caf3dc12eb72c4dea7477