Add markers to headings of highlighted error formatter.
Summary:
Adds markers to headings and prepares for adding markers to prefix of contextual code, by calculating the max width of the column based on the number of markers for a given position. This involves figuring out unique positions. This also involved refactoring a few minor functions to handle this and future changes.

In particular, the test expected files should now have [N] appended to the claim/reason headers. Additionally, the col width is now set to the correct max width for all the context lines; in a later diff, the newly added empty space will be filled with text like " [1,3] " for the markers.

Reviewed By: Wilfred

Differential Revision: D22452609

fbshipit-source-id: d96c1e99a75f64be6488baa0296f177a494fe6cc