Remove pseudomain for fatal units
Summary:
This diff is a step towards killing pseudomains.
When there's a parse error, we emit `String err` + `Fatal Parse` bytecodes to the pseudomain of the unit. D22443764 (https://github.com/facebook/hhvm/commit/1505934dea2779410b0b997b0e4b571c7668c3fa) started emitting a `.fatal` directive. This diff takes advantage of that and introduces `UnitFatal*` in addition to the existing `Unit*` in order not inflate the side of `Unit*` for all units and during merge we emit the parse/runtime fatal.

Reviewed By: ricklavoie

Differential Revision: D22465038

fbshipit-source-id: f4bbd936f1cfc93a84021f75a4c13996547fbfb0