Ban backtick from being parsed as a string
Summary:
Backticks are currently being lexed equivalently to double quotes. The lexer even returns the same token kind to the parser: either `TokenKind::DoubleQuotedStringLiteral` or `TokenKind::DoubleQuotedStringLiteralHead`.

This diff removes backtick from being lexed like double quote. We could throw this error after we've constructed the parse tree - but that requires us creating a new TokenKind for backtick or inspecting the text of the token in the parser. Neither solution seems like a good fit to simply throwing an error when we see backticks.

Reviewed By: Wilfred

Differential Revision: D22562252

fbshipit-source-id: fdcdbf45816c3d6e2ccb3f02510252c5c8ea9bc3