Return OpaqueValue from to_ocamlrep
Summary:
The values we store in cachelib are ocamlrep Slabs--contiguous, movable byte strings, each containing a tree of OCaml values. We currently write a decl to cachelib by allocating an OCaml value in an ocamlrep::Arena via ToOcamlRep, then copying that value into a Slab byte string by traversing the value tree (in `ocamlrep::slab::OwnedSlab::from_value`), then memcpying that byte string into cachelib-managed memory. The tree traversal we use to construct the slab is expensive--I've measured it taking as long as or longer than the ToOcamlRep conversion took in the first place (both conversion-to-OCaml and the slab-building traversal take roughly 25% of the time it takes to parse the decl). It would be much cheaper if the Arena built a slab in the first place--then we could skip the traversal and simply memcpy the slab into cachelib.

This wouldn't be so difficult if we knew ahead of time how much memory was needed for each decl. We could allocate the Arena with a sufficient initial capacity, build the OCaml value, extract the Vec backing the Arena's single chunk of memory, and use its contents as a slab. But we don't know how much memory will be needed for a decl before parsing it. It would be prohibitively expensive to use the maximum size of a decl for each arena--some files have a decl as large as 150MB, while the average file decl is only 12KB. We wouldn't want 24 decl-parsing threads each to hold on to 150MB of mostly-unused memory. Besides, our largest files may grow over time and eventually exceed whatever limit we choose.

Here's an alternative, though. We could write an Allocator implementation which uses a single Vec as its backing store (instead of a linked list of Vec-backed Chunks, as ocamlrep::Arena is implemented). Instead of returning pointers from allocation methods, it could return offsets into its backing Vec. That way, we'd guarantee that the value tree is contained in a single contiguous chunk of memory, making it essentially free to reinterpret that memory as a slab. That way, moving the slab into cachelib can be a simple memcpy.

However, ToOcamlRep::to_ocamlrep currently returns `ocamlrep::Value`s. These are inspectable OCaml values, which means they must either be an immediate integer or a pointer to a real memory address containing an OCaml block. If we want Allocators to have the freedom to return offsets instead, we need to use a non-inspectable type. We already have one in `ocamlrep::OpaqueValue` (a type added to implement Slabs, where the contained values may either be pointers or offsets). All we need to do to support our new use case is return OpaqueValues from to_ocamlrep instead of Values.

Reviewed By: arxanas

Differential Revision: D22361503

fbshipit-source-id: 927658cf3691bc5b29c4ef155dfc1464710bb549