support information flow for `new` expressions
Summary:
It is convenient to be able to use `new` in test cases, so I'm adding support for it.

The basic idea is to treat `new ClassName($a, $b, $c, ...)` as a call to the function `ClassName#__construct`. The one additional complexity is that constructors are assigned the type void, whereas we want to return a value with the type of the object. So, instead of just calling the construct function, we also have to create the returned `ptype` separately

Reviewed By: madgen

Differential Revision: D22516096

fbshipit-source-id: f46013d65681547b03a8ecaab97fb4078e12fdeb