Patch object type monotonicity
Summary:
The HHBBC type system isn't monotonic. It's possible to construct types a, b, A, and B, such that A <= a and B <= b but !(A|B <= a|b). In particular, if a and b are some interface type, and A and B are unrelated classes that both implement it, we'll hit this issue.

This bug causes us to produce non-monotonic type inferences: when two function return types get sharper, the type of the overall result gets less sharp. This behavior can occur even if we don't store ANY interface types in the index; we can do the non-monotonic update entirely within a function.

The proper way to fix this bug is to support much more precise specialized object types: for any given object, we can track an (exact or sub) class and *any number of interfaces*. If we intersect unrelated classes, we can still produce a specialized object with some list of known-to-be-implemented interfaces. However, that fix is highly nontrivial; a lot of logic will have to change to be executed multiple times rather than once.

Instead, we avoid loosening interfaces except when storing types in the index, at which point we loosen all interfaces. This diff addresses the common case where we run into this problem: functions or properties with an interface typehint which have many class types satisfying them.

Reviewed By: ricklavoie

Differential Revision: D22077957

fbshipit-source-id: 4fdea9d04b4f875887d0669aa3638dc5c0c638e8