Add CompactToken and CompactTrivia types and use them in direct decl parser
Summary:
In {D8119805 (https://github.com/facebook/hhvm/commit/ed55b0af0565b461547a0fb633bd3a42b25ae6f3)}, jewelpit traded a 3% parse time regression for a 3.8% memory savings by replacing the two `PositionedTrivia.t list`s on each positioned token with bitflags. While the change didn't reduce allocations (the lexer still produced the trivia lists--they were only collapsed into bitflags upon constructing a token), it ensured those lists were short-lived.

Because trivia is not used in the direct decl parser (except to determine the file mode specified in the comment trailing the `<?hh` token), and it is used only in simple ways within the FFP (e.g., "does this node have trailing trivia of kind WhiteSpace")?, I figured I could turn Julia's idea into a perf *improvement* for the direct decl parser. The regression came (in part) from needing to rescan trivia in circumstances where the FFP looked at the leading or trailing trivia of a token. But the FFP doesn't specifically need a trivia list--everything it is interested in knowing about a token's trivia can be answered from a bitflags type. We just need to revise the LexableToken trait so that it can be implemented cheaply by a token containing trivia bitflags.

We can take the perf improvement even further by avoiding allocating the trivia lists in the lexer in the first place. To do this, I renamed the LexableTrivia trait, which was formerly implemented by the data structure representing a single item of trivia, to LexableTrivium. I reused the plural name, LexableTrivia, for a new trait which represents the entire collection of trivia proceeding or following a token (i.e., a trivia list), and implemented it for Vec (for existing token types) and for trivia bitflags (for my new token type). The lexer can then construct whatever collection is appropriate, avoiding heap allocations in the bitflags case.

To further reduce heap allocations, I wanted to unbox tokens themselves. Trivia bitflags made this possible--since the bitflags type is only 1 byte, and token/trivia widths are generally very small, I was able to pack a positioned token representation in only 12 bytes (our current PositionedToken representation is an 88-byte struct behind an Rc). With a 4-byte-aligned enum tag (indicating whether the token is the small, inlined variant, or the large, boxed variant which uses usizes to represent widths that don't fit into the small variant), the entire token is only 16 bytes, the same size as a slice.

Reviewed By: shiqicao

Differential Revision: D22009525

fbshipit-source-id: 3aaf3f5729fb5b00da9e2bfb33d215b5b61601a9