Eliminate ArrayData::plusEq
Summary:
Another PHPism down. This operation wasn't HAM-safe, and it was a little nuts to boot.

There are two places that C++ code invoked this operator. In the debugger, we maintain the behavior by manual iteration. In ext_soap, we're listing available functions; there's a debugger command that does the same thing and assumes that system and user functions are disjoint, so maybe we can assume that too? Alternatively, I could just raise instead.

Reviewed By: billf, dneiter

Differential Revision: D22627817

fbshipit-source-id: 5a7a3ab3e097fcec8464301905e5d3c175999f48