Run Unit::merge until convergence
Summary:
This diff is a big step towards removing pseudomains. In order to completely avoid running pseudomains, we need to mark every unit as merge only. Prior to this diff only hhbbc would mark units as merge only since it would walk through the pseudomain and compute the correct order to merge the units however we will soon no longer have a pseudomain to walk through. This means that we need to change our merging strategy.

This diff introduces merging until convergence or a define time fatal.

This type of merging is a little more permissive than what we currently have but that's a step in the right direction since with symbol level loading this sort of permissiveness will be the new view.

After this diff, pseudomains can be completely removed.

Reviewed By: jano

Differential Revision: D22558752

fbshipit-source-id: 66d312ad3d06795228416a7bc45a94c6cee1884b