Highlighted error formatter edge cases: position outside file range, zero width positions at various columns
Summary:
Updates the highlighted error formatter to handle case where the error position is outside the range of the file. It currently prints `_` at the position outside the file range (whereas in the normal contextual error formatter, it would have put a caret `^` under the position in the next line).

It also fixes the algorithm that detects if a line or line/col pair are part of a position to handle several more edge cases related to zero-width positions and whether they occur at the beginning, middle, or end of a line.

My general approach to determining what to highlight is to iterate over each character in each line and highlight it if there is a corresponding marker for that position on the line. The approach would normally fail with error positions outside the file range, since they don't have a corresponding physical position for me to iterate over and see by default. To overcome this, I add a line at the end of the file if a non-zero-length position references that line, and I add a blank column at the end of the last line of a position if it references that column so that a out-of-bounds character can be added if needed.

**Example 1**
Before:

{F244983286}

After:

{F244387776}

**Example 2**
Before:

{F244983350}

After:

{F244387777}

**Example 3**
Before:

{F244983382}

After:

{F244387775}

**Example 4**
Before:

{F244983449}

{F244387774}

Reviewed By: Wilfred

Differential Revision: D22522616

fbshipit-source-id: 9946a470750466bc595758230d1583d262779f10