Mark Collections::map PSF
Summary:
In cases where we're mapping over a collection, and we have a  callback that returns a PHP array, if the callback is marked PSF then we also want `map` to be marked PSF as well so provenance flows up to the actual `map` call and not inside it where the callback gets executed.

This is done for Map/Pair/Vector since those have callbacks that could return arrays, but not for Set since those can't hold non-arraykey types.

Reviewed By: hgoldstein

Differential Revision: D22603880

fbshipit-source-id: 76e249d487078120814a3ae2d2aec40d47fb64ee