Ban meth_caller with traits
Summary:
Inside HHVM, we've adjusted the native `meth_caller` implementation to resolve the target class at time of construction. We should report an error if an attempt is made to refer to a method on a trait, since the resulting pointer is unusable anyway: we'll expect the object to be an "instance" of the trait, but traits are not types and there are no instances of them.

An example is below:

```
trait MyTrait {
  public function foo(): void {}
}

// Should throw an error, but does not currently
meth_caller(MyTrait::class, 'foo');
```

Reviewed By: Wilfred

Differential Revision: D22556864

fbshipit-source-id: 1e5cc7258079b3f762cadbf4696df7414409cf88