Ban many SPL iterators in HHI
Summary:
We'd like to delete methods of ArrayIterator. These methods are definitely not used in our codebase, but some of them are used to implement other SPL iterators (which are also unused in our codebase). Modify the HHI to hide all of these iterators prior to a runtime change to delete them. Deleted:

-  All methods of ArrayIterator other than current, key, next, rewind, valid
- CallbackFilterIterator, RecursiveCallbackFilterIterator
- CachingIterator, RecursiveCachingIterator
- RecursiveArrayIterator, RecursiveTreeIterator
- AppendIterator
- LimitIterator
- ParentIterator

Reviewed By: rodmk

Differential Revision: D22665716

fbshipit-source-id: f5809d01216df5b205db234c5f25306e8df079fa