Implicit class to string conversions
Summary:
This diff covers various places where a class pointer needs to be converted to a string because it is to be used as a key. As classes are used as keys pretty commonly, we want to avoid punting for class keys. Instead, we want to load the class name at JIT time. This is achieved by calling `convertClassKey` on keys. Similarly we call `tvClassToString` in interpreter. Consequently, we do not need to handle classes in `tvToKey`. Additionally we need to convert classes to strings for `SetArray`s.
As at the bytecode level we are now allowing classes to be used as keys, we also introduces a new type `TArrKeyLike` in HHBBC which is union of `TArrKey` and `TCls`.

Reviewed By: paulbiss

Differential Revision: D22505142

fbshipit-source-id: 50b65a79aa0424ba8dd1141c7a7d28f676d4cda4