initial support for information flow in catch
Summary:
In this diff, I add the initial support for information flow in try-catch.

To do this properly, we will need to use exception subtyping information which is quite complicated. I will leave that for a later diff.

For now, the behavior of of catch is quite similar to `if`, but instead of the pc being influenced by the scrutinee, it is now influenced by which exceptions are thrown. In addition, the local id that the exception is bound to also contains information about the excpetion itself.

Reviewed By: madgen

Differential Revision: D22526539

fbshipit-source-id: a4435156a1734524e4d039f2ec7736a0aa25dcd4