Eliminate plain arrays
Summary:
In this change, we make it impossible to create a "plain" (non-dvarray) PHP array at runtime. The main place it's possible to create such arrays is in mixed-array.cpp, when we instantiate an array with HeaderKind::Plain. We delete these code paths. We also delete the ToPHPArray ArrayKind interface function; we simply reroute any calls to this function to ToDArray instead. Similarly, we reroute staticEmptyArray and ArrayData::Create to their darray counterparts, too.

After we land this change, we'll want to do some cleanups where we replace usages of ArrayInit and MixedArrayInit with DArrayInit, delete KindOfArray, delete HHBBC's PArr and the JIT's PArr, etc. But those are all simple refactorings that we can do as follow-ups. They'll result in a large amount of deleted code, but no more behavior changes.

Safety:

This change looks totally unsafe! In fact, it should be safe to a Hack codebase that's been prepared correctly, as ours has. Right now, plain arrays can only ever be constructed by C++ code within builtins. The key criteria we need to make this change safe is the "barrier" criterion: **no builtin ever returns a plain array result to Hack code**. One way to ensure this criterion - the way we've taken in our codebase - is to wrap all builtins in a wrapper (perhaps "darray_recursive" or "darray_shallow") that ensures that any plain arrays in the result are converted into darrays.

Why is this criterion sufficient? Note that there are very few behavior differences between plain arrays and darrays. darrays can do weird things that plain arrays can. They cause builtins to intish-cast keys, they support append syntax, and they even have the wacky unset-append behavior that PHP arrays do. Here's a list of the actual differences:
1. Plain arrays don't compare equal/same to darrays with equal/same elements.
1. Plain arrays support relational comparison with <, <=, >, >=; darrays throw.
1. darray typehints allow darrays but reject plain arrays (plain array typehints are dead)

(This list is a subset of the behaviors we changed for HAM specialization, since before that landed, there was no behavior difference between plain arrays, varrays, and darrays. The other specialization behaviors don't apply; one of them was stricter checks on class-subclass types, and the other one was much tighter restrictions on varray behaviors.)

Builtins don't use the deep comparison operators (equality and relational); they're only accessed from Hack code. As a result, the barrier criterion ensures that first two behavior differences go unobserved. The same criterion ensures that this change will never result in typehint violations, because any value checked against a typehint is a value exposed to Hack, which can't be a plain array by the barrier criterion.

Reviewed By: jano

Differential Revision: D22365647

fbshipit-source-id: 93b2dac7afd84f9d10d77e13916303f783c0e60a