Simplify ocaml_to_isize
Summary: To convert an OCaml int to an isize, you just need to perform an arithmetic right shift by one bit. Since the shift-right operator performs a *logical* right shift on usizes, this function does some bit masking to get the behavior of an arithmetic shift. But there's no need--the shift-right operator performs an arithmetic shift on isizes, so we just need to cast first.

Reviewed By: andrewjkennedy

Differential Revision: D22456001

fbshipit-source-id: 1eae1529f6ab8cab114622117ed70093f515ca12