Add self parameter to Arena::set_field
Summary:
The set_field method was originally added to the Allocator trait in anticipation of the change in this stack--we'd like to support using offsets into a Vec as values instead of pointers, so that we can cheaply produce slabs (i.e., with a memcpy instead of a tree traversal). There was an oversight, though--if set_field is provided with an offset, but no reference to the allocator, then there is no way to locate the field in memory. We cannot have a Vec-based arena return BlockBuilders containing actual addresses, since the Vec may be reallocated between the time the BlockBuilder was created and a field was set. As a result, set_field needs the allocator in scope, so that we can index into the backing Vec and set the field in the block at the BlockBuilder's offset.

This diff only adds a self parameter to set_field in anticipation of that change.

Reviewed By: arxanas

Differential Revision: D22361506

fbshipit-source-id: 26b4616d96dcd1f5c7a464ef2eef21cd93afdb3a