Be more specific about when MaybeUninit is necessary in slab module
Summary: The use of `MaybeUninit` was added in a codemod, after the `bytes` crate was changed to make use of it. Since cachelib exposes cache items using `bytes`, and we copy slabs into cache items, the change propagated into our slab module. But we don't need to handle the MaybeUninit case most of the time--only in the `copy_slab` function.

Reviewed By: arxanas

Differential Revision: D22361504

fbshipit-source-id: b07173e96e9653b0fa2b7d07341cc1668972a69a