Use the same marker/color for identical positions in highlighted error formatter
Summary: For positions that are identical in an error, use the same marker/color for highlighting them. This helps eliminate the problem (partially solved in diffs higher in the stack) of differentiating overlapping positions on the same line. For these cases, there is no need to differentiate them since they are the same.

Reviewed By: Wilfred

Differential Revision: D22636080

fbshipit-source-id: a94ce5ae1a42755d7797948cbd7bece5adec0863