Using text color, add a way to distinguish positions that are within another position in highlighted error formatter.
Summary:
Currently when two positions overlap, if the lower-numbered (i.e. probably more important one) overlaps with the other, only the color of the lower-numbered position is used.

Now, choose the text color of the smaller (in length) position so the boundaries between overlapping positions are visible. When there are ties, take the color of the marker with the lowest-number.

Reviewed By: Wilfred

Differential Revision: D22612696

fbshipit-source-id: 59b775be4174faf24b9c7d3093a5768436fec8a6