ban array typehints in the typechecker by default
Summary:
We currently have runtime errors with no line numbers, and sometimes the wrong file. We need the typechecker support; also, banning in typechecker should generally happen before the runtime.

- flip the default
- remove `toArray()`/`fromArray()` functions from HHI if there is varray/darray versions. Most of these have also already been removed from the runtime
- add typehints from `PHP\` if safe; otherwise, fall back to PHP docs or the most generic thing
  - if changing something to return `varray`, inspected C++ source to verify that was actually consistently happening
- delete FB-only `<<__PHPStdLib>>` definitions
- delete FB-only files that are now empty
- started manually updating tests (needed to be manual to preserve test indent); way too many to finish the lot in a reasonable timeframe for release; gave up on doing that, and flipped the default in hh_single_type_check to not match hh_server; remaining test changes are line numbers in HHI files

This diff also adds a `disable_parse_array_typehint` .hhconfig option - this should not be used in real projects: runtime support for array typehints is not just disabled - it's pretty broken in master if explicitly enabled.

This option has been added purely to allow a gradual migration of the typechecker's own tests.

Differential Revision: D22775331

fbshipit-source-id: 234a78136b10f2bec1d3f127e6b04e453343b725